# swagger.yaml
openapi: 3.0.3
info:
  title: Backend prova esame - Gestire delle consegne di un corriere espresso
  description: API documentation generated from Fastify route schemas
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /api/clienti/:
    get:
      tags:
        - Clienti
      responses:
        "200":
          description: Default Response
    post:
      tags:
        - Clienti
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - Nome
                - Via
                - Comune
                - Provincia
              properties:
                Nominativo:
                  type: string
                Via:
                  type: string
                Comune:
                  type: string
                Provincia:
                  type: string
                Telefono:
                  type: string
                Email:
                  type: string
                  format: email
                Note:
                  type: string
      responses:
        "200":
          description: Default Response
  /api/clienti/{id}:
    get:
      tags:
        - Clienti
      parameters:
        - schema:
            type: integer
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
    put:
      tags:
        - Clienti
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Nominativo:
                  type: string
                Via:
                  type: string
                Comune:
                  type: string
                Provincia:
                  type: string
                Telefono:
                  type: string
                Email:
                  type: string
                  format: email
                Note:
                  type: string
      parameters:
        - schema:
            type: integer
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
    delete:
      tags:
        - Clienti
      parameters:
        - schema:
            type: integer
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
...
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []